Accept Time objects as bind params
SQLite has 'datetime' columns, so it seems natural to pass Time objects in as bind
params when executing a query.

It would also be good to support Date and DateTime objects (from 'datetime' in
stdlib) in the future. These are not built-in Ruby core classes, so there is no
global variable like 'rb_cTime' which can be used to refer to them.