fix invalid memory access
As I already wrote in an earlier commit message, the memory access
violation occurs after the library user installs a second aggregate
handler. In that case, Database.@aggregate is overwritten and
therefore no references to the AggregateHandler remain. The simplest
fix would be to change @aggregator (or @agregator) into a Ruby array of
all AggregateHandlers ever installed. I decided against that because I
think its cleaner when SQLite controls the lifetime of the handlers. To
that end the sqlite developers extended the sqlite3_create_function API
with the xDestroy callback. The extended "v2" Version is available
since 3.7.3 (2010-10-08). I installed a fallback to the old version,
should the new function be unavailable. While it is possible to remove
elements from an @aggregator ruby array while in xDestroy callback,
this was clumsy and difficult to do correctly under undefined
finalization order. I came to the conclusion that its easier to manage
the lifetime completely in C and only provide the appropriate
rb_gc_mark function to SQLite3::Database. The remaining issue is to
make the VALUEs for the AggregateHandlers (and later their instances)
visible to the mark function. To that end, I introduced the
aggregator_wrapper struct that holds the VALUE and the linked list
pointers. I replaced @aggregator with the _sqlite3Ruby.aggregators
linked list head.

I used valgrind to check that memory accesses are sane now. Before this
patch, I saw loads from previously free'd memory that disappeared.
There are still lots of warnings about use initialized values, but this
is unrelated and most likely an internal issue with MRI 2.0.0.

As part of this patch, I moved the aggregate function related code into
its own files "aggregator.[ch]".