use sqlite3_aggregate_context()
This commit addresses the issue that using the same aggregate function
twice in a select statement mixes up the execution contexts in the
AggregateHandler. While resetting the aggregate Handler in the
finalize() method emulates the context lifetime in the other cases
(group by, independent select), resetting is a fragile approach.
Instead sqlite3_aggregate_context() is the function in the SQLite API
that should be used to separate contexts.

sqlite3_create_context() is a bit weird though: it is memory allocator
that returns the same memory when called from a context that already
called sqlite3_create_context once. Equally strange is that there is no
destroy() callback -- sqlite will clear up the memory when it does not
need it any more and won't tell us. Because a second allocator is
difficult to combine with Ruby's GC and because I do not have the
lifetime control, I use it only to store a pointer to a heap allocated
aggregator_instance_t structure. This structure holds the VALUE for the
Ruby instance that belongs to the aggregate function's execution
context. The structure forms a linked list rooted in the
aggregate_wrapper_t for the function it as an instance of, so the
instance can be GC marked.

SQLite's API left me with no choice but to assume that every execution
context ends with a single call to finalize(). At that point I unlink
and destroy the aggregate_instance_t. I never observed SQLite doing
anything else. There are checks in place to catch if SQLite does call
finalize() a second time. If it does not ever call finalize(), we leak
memory. If someone wants to use sqlite3_result_error, he or she should
check if the invocation of sqlite3_result_error is also the right place
to destroy the aggregate_instance_t.

This commit comes with one rather big change: now sqlite3-ruby
allocates the instances of the aggregate handler in the C code. This
forced me to change the interface, because the existing interface
passed only the instance object, not the factory, to the C code. To
that end, I introduced a new private method Database.define_aggregator2
that takes the factory as it sole parameter. Its like
Database.create_aggregate_handler but without the FunctionProxy "ctx"
object that serves no useful purpose. I wrote a new Ruby shim for the
old Database.define_aggregator and adopted Database.create_aggregate
and Database.create_aggregate_handler to the new interface between Ruby
and C. Generally I was careful to keep the interface as it is.

There is a tiny API change that I could not help avoid:
As Database.define_aggregator only receives one object, but it needs
one per execution context, the only option i saw was to use the
provided one as template and give every new instance a clone. This
means, however, that the template must provide a suitable clone. As
every ordinary ruby object already provides one, this should not pose
problems in practice, though.

And of course, people who were getting wrong results from their
aggregate functions suddenly get the right one, which may also come as
a surprise to them.