Add a `strict: true` option.
Ref: https://github.com/rails/rails/pull/45101
Ref: https://github.com/rails/rails/issues/27782
Ref: https://www.sqlite.org/quirks.html#double_quoted_string_literals_are_accepted

Sqlite has a really unfortunate behavior around double quotes. It first
tries to consider them as column names, but if they don't exist it then consider
them as string literals.

Because of this, typos can silently go unnoticed. It would be very useful
for Active Record if this behavior could be disabled.