Use a factory method to eliminate "results as hash" check
"Results as hash" should be set on the database, so we should only need
to check it once when we first start the query.  This patch pulls result
set construction in to a factory method on the database object.  Doing
this eliminates a "translate to hash" check that is performed on every
row fetch.