feat: Database.new supports `extensions:` option parameter
First change: Database#load_extension supports an extension
specifier, which is anything that responds to `#to_path` like Pathname
or the SQLean gem modules.

When passed an object that responds to `#to_path`, that method's
return value is used as the filesystem path for the extension,
allowing a calling convention like:

    db.load_extension(SQLean::VSV)

where prior to this change that would need to be the more verbose:

    db.load_extension(SQLean::VSV.to_path)

Second change: Database.new supports `extensions:` option parameter,
which is an array of paths or extension specifiers to be loaded.

This commit updates the Database documentation with lots of examples
and explanation, but to sum up:

   Database.new(":memory:", extensions: [SQLean::Crypto, "/path/to/extension"])

My hope is that we can now make a very small change to Rails to allow
injection of database extensions via the config/database.yml file
using this constructor param.